凍った時の見た目変更

+++ sugiEngine/app/enemy/BaseEnemy.h
@@ -19,7 +19,7 @@ struct DebuffM {
 
 struct DownState {
 	float slow = 1.0f;
-	int damage = 0;
+	float damage = 0;
 };
 
 class BaseEnemy {
@@ -90,7 +90,7 @@ public:
 	void SetIsStop() {
 		isStop_ = true;
 	}
-	uint32_t GetLife() {
+	float GetLife() {
 		return life_;
 	}
 
@@ -123,7 +123,7 @@ public:
 	/// </summary>
 	/// <param name="subLife">hp減少量</param>]
 	/// <param name="subLife">パーティクルを出すか</param>
-	void SetIsHit(int32_t subLife, bool isParticle = false);
+	void SetIsHit(float subLife, bool isParticle = false);
 
 	/// <summary>
 	/// デバフ状態になっているか
@@ -175,7 +175,7 @@ protected:
 	/// </summary>
 	/// <param name="subLife">減らすhp量</param>
 	/// <param name="subLife">パーティクルを出すかどうか</param>
-	void SubLife(int32_t subLife,bool isParticle = false);
+	void SubLife(float subLife,bool isParticle = false);
 
 	// デバフの効果を更新
 	void UpdateDebuff();
@@ -197,7 +197,7 @@ protected:
 	//何フレームごとにバーンダメージをくらうか
 	const int32_t TIME_FIRE_BURN = 40;
 	//一回のバーンダメージでくらうダメージ量
-	const int32_t DAMAGE_FIRE_BURN = 1;
+	const float DAMAGE_FIRE_BURN = 1;
 	const float MAX_SHAKE = 0.5f;
 	//シェイク量調整
 	const float CALC_SHAKE = 100;
@@ -216,7 +216,7 @@ protected:
 	float height_;
 
 	//体力
-	int32_t life_;
+	float life_;
 
 	//死んだかどうか
 	bool isDead_;
